use floating point precision consistently

+++ ThrEd4/satin.h
@@ -56,18 +56,18 @@ namespace internal {
 
 	uint32_t nusac(uint32_t formIndex, uint32_t guideCount);
 
-	void     outfn(uint32_t start, uint32_t finish, double satinWidth);
+	void     outfn(uint32_t start, uint32_t finish, float satinWidth);
 	void     sacspac(uint32_t startGuide, uint32_t guideCount);
 	void     satclos();
 	void     satcpy(const std::vector<SATCON>& source, uint32_t size);
 	void     satends(uint32_t isBlunt);
-	void     satfn(const std::vector<double>& lengths,
+	void     satfn(const std::vector<float>& lengths,
 	               uint32_t               line1Start,
 	               uint32_t               line1End,
 	               uint32_t               line2Start,
 	               uint32_t               line2End);
 	uint32_t satind(const SATCON* guide) noexcept;
-	void     satmf(const std::vector<double>& lengths);
+	void     satmf(const std::vector<float>& lengths);
 	void     satsbrd();
 	bool     satselfn();
 	void     sbfn(const std::vector<fPOINT>& insidePoints, uint32_t start, uint32_t finish);
